+++ include/vec/crud.h
@@ -6,6 +6,7 @@
 
 Vec vec_new ();
 Vec vec_new_n (long capacity);
+VecNamespace vec_free (Vec v);
 Vec vec_delete (Vec v);
 
 long vec_length (Vec v);