+++ HISTORY.md
@@ -10,7 +10,8 @@ X.Y.Z | (goals)
   * Added `cursor`
   * Implemented `vec_length()` and `vec_capacity()` functions
   * Implemented `vec_empty()` and `vec_full()` functions
-  * Added magic method fields to Vec: `less_than()`, `hash_code()` and `to_string()`
+  * Added magic method fields to `Vec`: `less_than()`, `hash_code()` and `to_string()`
+  * Added magic method field `clone()` to allow for creating a duplicate copy of `Vec`
 
 
 0.1.0 | 07-11-2020